+++ lib/flipflop/feature_set.rb
@@ -28,7 +28,10 @@ module Flipflop
       private :new
     end
 
+    attr_reader :raise_strategy_errors
+
     def initialize
+      @raise_strategy_errors = (ENV["RACK_ENV"] || ENV["RAILS_ENV"]) != "test"
       @features = {}
       @strategies = {}
     end